ENG-22363 use round-robin for compound procs (#1247)
* make compound test work better with round-robin

* use round-robin for compound procs

+++ tests/test_apps/compound/ddl.sql
@@ -11,7 +11,7 @@ create table foo (intval integer not null, strval varchar(20), primary key(intva
 partition table foo on column intval;
 create table bar (intval integer not null, strval varchar(20), primary key(intval));
 partition table bar on column intval;
-create table mumble (intval integer not null, strval1 varchar(20), strval2 varchar(20), primary key(intval));
+create table mumble (intval bigint not null, strval1 varchar(20), strval2 varchar(20), primary key(intval));
 partition table mumble on column intval;
 create procedure MySpProc partition on table foo column intval as select strval from foo where intval = ?;
 create procedure MyOtherSpProc partition on table bar column intval as select strval from bar where intval = ?;
